Extract DoWhile to separate struct
